Refactor helpers -> utils
