+++ ocrd_network/ocrd_network/param_validators.py
@@ -1,6 +1,6 @@
 from click import ParamType
 
-from ocrd_network.helpers import (
+from ocrd_network.utils import (
     verify_database_uri,
     verify_and_parse_mq_uri
 )
